Move demo.py into the Python examples as quickstart.py
The root `examples/` directory held a single file that nothing linked
to and no CI job ran. Its API calls were all still valid, but the
enqueue and claim halves of the quickstart were otherwise split across
`atomic.py` (which never claims) and `worker.py` (which never enqueues
inside a transaction), so the README quickstart had no runnable
counterpart.

Move it to `packages/honker/examples/quickstart.py` and list it first in
the examples README. Switch it to a `tempfile.TemporaryDirectory` like
its siblings: it previously wrote `app.db` into the caller's cwd and
cleaned up only the main file, leaving `-wal`/`-shm` behind, which
contradicted the README's claim that examples leave nothing behind.

Also assert on `job.ack()` and raise the worker timeout to 5s to match
the other examples.
